fix cctools installation and correct exec script in taskvine

+++ parsl/executors/taskvine/install-taskvine.sh
@@ -7,6 +7,10 @@ if [ "$CCTOOLS_VERSION" == "" ] ; then
   exit 1
 fi
 
-wget -O /tmp/cctools.tar.gz "https://github.com/cooperative-computing-lab/cctools/releases/download/release/$CCTOOLS_VERSION/cctools-$CCTOOLS_VERSION-x86_64-ubuntu20.04.tar.gz"
+if [ -f "/etc/redhat-release" ]; then
+    wget -O /tmp/cctools.tar.gz "https://github.com/cooperative-computing-lab/cctools/releases/download/release/$CCTOOLS_VERSION/cctools-$CCTOOLS_VERSION-x86_64-centos8.tar.gz"
+else    
+    wget -O /tmp/cctools.tar.gz "https://github.com/cooperative-computing-lab/cctools/releases/download/release/$CCTOOLS_VERSION/cctools-$CCTOOLS_VERSION-x86_64-ubuntu20.04.tar.gz"
+fi
 mkdir -p /tmp/cctools
 tar -C /tmp/cctools -zxvf /tmp/cctools.tar.gz --strip-components=1
